[refactor] deplacer la table dans share [style] modifier la couleur du text dans le tableau et ajout de striped row
