+++ src/shared/Table.css
@@ -1,0 +1,3 @@
+.striped-row {
+    background-color: #c0bcbc;
+}
