+++ src/shared/index.js
@@ -2,5 +2,6 @@ import DynamicHeadNav from './DynamicHeadNav'
 import Text from './Text'
 import Row from './Row'
 import Column from './Column'
+import TableStriped from './TableStriped'
 
-export { DynamicHeadNav, Text, Row, Column }
+export { DynamicHeadNav, Text, Row, Column, TableStriped }
